// src/constants/enums.ts
export const enum USER_STATUS {
  ACTIVE = 0,
  INVITED = 1,
  INACTIVE = 2,
  BLOCKED = 3,
}

export const enum USER_ACCESS_ADMIN_STATUS {
  ACTIVE = 0,
  INACTIVE = 1,
}

export enum BOOL {
  FALSE = 0,
  TRUE = 1,
}

export enum ServiceResponseStatus {
  SUCCESS = 'success',
  FAIL = 'fail',
}

export enum ENV_TYPE {
  DEVELOPMENT = 'DEVELOPMENT',
  STAGING = 'STAGING',
}